sales-staff-agent/pyproject.toml
[project]
name = "sales-staff-agent"
version = "0.1.0"
description = "販売店スタッフ支援 Hosted Agent"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.13.3",
    "azure-ai-projects>=1.0.0b1",
    "azure-identity>=1.15.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
